requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
python-dotenv==1.0.1
typing_extensions==4.12.2
watchdog==6.0.0
